+++ pole_position/template/src/app/api/router.py
@@ -1,13 +1,9 @@
 from fastapi import APIRouter
 
-from {{project_import_name}}.modules.profile.router import router as profile_router
-from {{project_import_name}}.modules.races.router import router as races_router
 from {{project_import_name}}.modules.status.router import router as status_router
 # polepos:router-imports
 
 
 api_router = APIRouter()
 api_router.include_router(status_router, tags=["status"])
-api_router.include_router(profile_router, prefix="/profile", tags=["profile"])
-api_router.include_router(races_router, prefix="/races", tags=["races"])
 # polepos:router-includes