+++ pole_position/template/src/app/db/models.py
@@ -1,3 +1,2 @@
 def import_models() -> None:
-    from {{project_import_name}}.modules.races import model  # noqa: F401
     # polepos:model-imports