+++ pole_position/template/src/app/modules/__init__.py
@@ -1,6 +1,4 @@
 __all__ = [
-    "profile",
-    "races",
     "status",
     # polepos:module-exports
 ]