+++ pole_position/template/src/app/modules/races/__init__.py
@@ -1,7 +1,0 @@
-__all__ = [
-    "model",
-    "repository",
-    "router",
-    "schemas",
-    "service",
-]